backend/requirements.txt
drf-yasg==1.21.7
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-oauth-toolkit==2.3.0
psycopg2-binary==2.9.10
environ==1.0
python-dotenv==1.0.0
pytest==8.3.5